network: add: look for remote by its 16-bit addr before adding if no 64-bit addr
If a remote to be added to the network does not have its 64-bit address defined
or it is unknown, the method tries to get it (params SH and SL). If this fails,
it now looks for it in the network by its 16-bit address. If found, that is the
node to be updated.

This may happen when generating the route from the local XBee to a remote in
Zigbee, if the intermediate hops are not in the network, because the 'Route
Record Indicator' frame (0xA1) only provides the 16-bit addresses of the hops.

https://jira.digi.com/browse/XBPL-333

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>